+++ CotEditor/Sources/EditorTextView.swift
@@ -55,7 +55,7 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, Multi
     var syntaxCompletionWords: [String] = []
     
     var needsUpdateLineHighlight = true
-    var lineHighLightRect: NSRect?
+    var lineHighLightRects: [NSRect] = []
     private(set) var lineHighLightColor: NSColor?
     
     var insertionLocations: [Int] = [] { didSet { self.updateInsertionPointTimer() } }
@@ -1445,6 +1445,10 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, Multi
         }()
         
         let replacementString = urls.reduce(into: "") { (string, url) in
+            if url.pathExtension == "textClipping", let textClipping = try? TextClipping(url: url) {
+                string += textClipping.string
+                return
+            }
             if let dropText = composer.dropText(forFileURL: url, documentURL: documentURL, syntaxStyle: syntaxStyle) {
                 string += dropText
                 return