+++ CotEditor/Sources/EditorTextViewController.swift
@@ -9,7 +9,7 @@
 //  ---------------------------------------------------------------------------
 //
 //  © 2004-2007 nakamuxu
-//  © 2014-2018 1024jp
+//  © 2014-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -65,6 +65,7 @@ final class EditorTextViewController: NSViewController, NSTextViewDelegate {
         super.viewDidLoad()
         
         // observe text orientation for line number view
+        self.orientationObserver?.invalidate()
         self.orientationObserver = self.textView!.observe(\.layoutOrientation, options: .initial) { [unowned self] (textView, _) in
             
             switch textView.layoutOrientation {