+++ CotEditor/Sources/FindPanelFieldViewController.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2014-2018 1024jp
+//  © 2014-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -60,6 +60,7 @@ final class FindPanelFieldViewController: NSViewController, NSTextViewDelegate {
         
         super.viewDidLoad()
         
+        self.defaultsObservers.forEach { $0.invalidate() }
         self.defaultsObservers += [
             // sync history menus with user default
             UserDefaults.standard.observe(key: .findHistory, options: .initial) { [unowned self] _ in
@@ -226,6 +227,8 @@ final class FindPanelFieldViewController: NSViewController, NSTextViewDelegate {
     
     /// apply history to UI
     private func buildHistoryMenu(_ menu: NSMenu, defaultsKey key: DefaultKey<[String]>, action: Selector) {
+        
+        assert(Thread.isMainThread)
         
         // clear current history items
         menu.items