+++ CotEditor/Sources/NSDocument+ErrorHandling.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2016-2018 1024jp
+//  © 2016-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -24,6 +24,22 @@
 //
 
 import AppKit.NSDocument
+
+extension NSDocument.SaveOperationType {
+    
+    var isAutoSaving: Bool {
+        
+        switch self {
+        case .autosaveElsewhereOperation, .autosaveInPlaceOperation, .autosaveAsOperation:
+            return true
+        case .saveOperation, .saveAsOperation, .saveToOperation:
+            return false
+        }
+    }
+    
+}
+
+
 
 extension NSDocument {